refactor(workbench): decompose host dock orchestration
