+++ packages/workbench/surface/src/host/dockActions.ts
@@ -1,0 +1,3 @@
+export function dockActionKey(entryId: string, actionId: string): string {
+  return `${entryId}:${actionId}`;
+}
